chore(pnpm): use pnpm instead of npm (#69)
#### What this PR does / why we need it:

With pnpm, it creates a [non-flat
node_modules](https://pnpm.io/motivation#creating-a-non-flat-node_modules-directory)
directory. When installing dependencies with npm or Yarn Classic, all
packages are hoisted to the root of the modules' directory. As a result,
the source code has access to dependencies that are not added as
dependencies to the project.

Hosting is an incredibly challenging problem in mono-repo projects, node
module resolver needs to know where to look for dependencies. When
project dependencies are not hoisted automatically due to version
conflicts (see below), it creates multiple copies of a dependency at
different layers of your package hierarchy, makes it impossible to
manage.

```text
├── package.json
├── services/
│   └── service-a
│       ├── AService
│       └── node_modules
│           └── @birthdaycloud/deps-b@1.0.0
└── node_modules
    └── @birthdaycloud/deps-a@1.0.0
```

Given the above, we're using the node default package resolution
technique, and `@foo/deps-a` depends on `@foo/deps-b` and you require
`@foo/deps-a` for `service-a`. Since `@foo/deps-b` is a dependency of
`@foo/deps-a` and it isn't intrinsically defined within
`service-a/package.json`, the default npm behavior would be to
deduplicate by hoisting it to the root `./node_modules`. The runtime
would panic as it won't know how to resolve `@foo/deps-b` from
`service-a/node_modules` since it's hoisted up while `@foo/deps-b`
isn't.


**Migration Plan:**

- [ ] Keep `ts-tests` isolated for now, while focusing on a clean pnpm
structure and migrating `ts-tests` into `packages/*` at a later date.